+++ src/me/theguyhere/villagerdefense/game/displays/Portal.java
@@ -70,6 +70,7 @@ public class Portal {
 
 	public void removePortalAll(int arena) {
 		removeNPCAll(arena);
+		NPC[arena] = null;
 		if (holos[arena] != null)
 			holos[arena].delete();
 	}
@@ -123,9 +124,11 @@ public class Portal {
 
 	public void refreshHolo(int arena, Game game) {
 		Arena arenaInstance = game.arenas.get(arena);
-		holos[arena].delete();
+		if (holos[arena] != null)
+			holos[arena].delete();
 		Location location = utils.getConfigLocationNoPitch("portal." + arena);
-		addHolo(location, arena, getHoloText(arenaInstance));
+		if (location != null)
+			addHolo(location, arena, getHoloText(arenaInstance));
 	}
 
 	public void removeAll() {
@@ -138,16 +141,42 @@ public class Portal {
 
 	private String[] getHoloText(Arena arena) {
 		String status;
-		if (arena.isClosed())
-			status = "&4&lClosed";
+
+		// Get difficulty
+		String difficulty = arena.getDifficultyLabel();
+		if (difficulty != null)
+			switch (difficulty) {
+				case "Easy":
+					difficulty = " &a&l[" + difficulty + "]";
+					break;
+				case "Medium":
+					difficulty = " &e&l[" + difficulty + "]";
+					break;
+				case "Hard":
+					difficulty = " &c&l[" + difficulty + "]";
+					break;
+				case "Insane":
+					difficulty = " &d&l[" + difficulty + "]";
+					break;
+				default:
+					difficulty = "";
+			}
+		else difficulty = "";
+
+		// Get status
+		if (arena.isClosed()) {
+			return new String[]{Utils.format("&6&l" + arena.getName() + difficulty),
+					Utils.format("&4&lClosed")};
+		}
 		else if (arena.isEnding())
 			status = "&c&lEnding";
 		else if (!arena.isActive())
 			status = "&5&lWaiting";
 		else status = "&a&lWave: " + arena.getCurrentWave();
-		return new String[]{Utils.format("&6&l" + arena.getName()),
+
+		return new String[]{Utils.format("&6&l" + arena.getName() + difficulty),
 		Utils.format("&bPlayers: " + arena.getActiveCount() + '/' + arena.getMaxPlayers()),
-		Utils.format("&7Spectators: " + arena.getSpectatorCount()),
+		Utils.format("Spectators: " + arena.getSpectatorCount()),
 		Utils.format(status)};
 	}
 }
